Make all docker images consistent for future updates (#35919)
We have update CLI that takes care of Docker image updates. In order
to cover all the images we need to make them consistent first.

If we keep these images at older versions, we'll be constantly
receiving CVE notifications, so it's easier to automate these updates
even though these images are internal and are never released to the public.